퀴즈채점, 퀴즈셋 받아오기

+++ backend/models/Quiz.go
@@ -10,6 +10,6 @@ type Quiz struct {
 	QuizTitle string `db:"quiz_title" json:"quiz_title"`
 	QuizContent string `db:"quiz_content" json:"quiz_content"`
 	QuizType string `db:"quiz_type" json:"quiz_type"`
-	QuizScore uint `db:"quiz_score" json:"quiz_score"`
+	QuizScore uint64 `db:"quiz_score" json:"quiz_score"`
 	QuizAnswer string `db:"quiz_answer" json:"quiz_answer"`
 }
